improved detection of which zone the user is in.
instead of using euclidian distance to the centeroit we now use a haversine distance to the closest exterior point. this way we can support zones that go acrous the 180th meridian/international date line. by switching to exterior points the detection is better in large overlapping zones as the centerpoint of the best zone coukld be furter than the worse fitting zone